+++ Changelog.rst
@@ -9,6 +9,8 @@ New features
 
 * Added support for Poly1305 MAC (with AES and ChaCha20 ciphers for key derivation).
 * Added support for ChaCha20-Poly1305 AEAD cipher.
+* New parameter ``output`` for ``Crypto.Util.strxor.strxor`` and ``Crypto.Util.strxor.strxor_c``
+  to have the result put into a pre-allocated buffer (e.g. a ``bytearray``).
 
 Breaks in compatibility
 -----------------------